pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-dataiku"
version = "0.1.0"
description = "MCP server for Dataiku - expose Dataiku DSS capabilities as MCP tools"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["mcp", "dataiku", "dss", "ai", "llm"]
dependencies = [
    "mcp>=1.0.0",
    "dataiku-api-client>=13.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
mcp-dataiku = "mcp_dataiku.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_dataiku"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
